assignment5a/tasks/requirements.txt
Flask==2.0.3
requests
google-cloud-storage
google-cloud-datastore
